pyproject.toml
[project]
name = "weekmenu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.9.2",
    "litestar[standard]>=2.16.0",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "tomlkit>=0.14.0",
    "tortoise-orm>=0.25.1",
]

[dependency-groups]
dev = [
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "ruff>=0.15.0",
]

[tool.aerich]
tortoise_orm = "src.db_config.TORTOISE_CONFIG"
location = "./migrations"
src_folder = "./."
